Changing to float

+++ Insurance/Customer.cs
@@ -8,25 +8,25 @@ namespace Insurance
 {
     public class Customer
     {
-        public double CalcPremium(int age, string gender)
+        public float CalcPremium(int age, string gender)
         {
-            double premium;
+            float premium;
 
             if (gender == "female")
             {
                 if ((age >= 18) && (age <= 30))
                 {
-                    premium = 5.0;
+                    premium = 5.0f;
                 }
                 else
                 {
                     if (age >= 31)
                     {
-                        premium = 2.5;
+                        premium = 2.5f;
                     }
                     else
                     {
-                        premium = 0.0;
+                        premium = 0.0f;
                     }
                 }
             }
@@ -34,27 +34,27 @@ namespace Insurance
             {
                 if ((age >= 18) && (age <= 35))
                 {
-                    premium = 6.0;
+                    premium = 6.0f;
                 }
                 else
                 {
                     if (age >= 36)
                     {
-                        premium = 5.0;
+                        premium = 5.0f;
                     }
                     else
                     {
-                        premium = 0.0;
+                        premium = 0.0f;
                     }
                 }
             }
             else
             {
-                premium = 0.0;
+                premium = 0.0f;
             }
             if (age >= 50)
             {
-                premium = premium * 0.15;
+                premium = premium * 0.15f;
             }
             return premium;
         }
